# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Produtos",
        "description": "Documentação da API de Produtos usando Swagger(antes de criar os produtos a(s) categoria precisam ser criada primeiro)",
        "contact": {
            "email": "suporte@exemplo.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar todos os produtos",
                "description": "Retorna a lista completa de produtos",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Lista de produtos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nome": {
                                                "type": "string",
                                                "example": "Camiseta"
                                            },
                                            "preco": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 49.99
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "novo do produto",
                "description": "Criação do produto",
                "operationId": "Criar um novo Produto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Camiseta"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição da camiseta"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 49.99
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "example": "9f2846be-da65-44f6-b453-6abdde55acb1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mostra o produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nome": {
                                                "type": "string",
                                                "example": "Camiseta"
                                            },
                                            "preco": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 49.99
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "category_id": {
                                                "type": "string",
                                                "example": "9f2846be-da65-44f6-b453-6abdde55acb1"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/product/{product}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "detalhes do produto",
                "description": "Detalhamento do produto",
                "operationId": "recuperar o produto por id",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID do Produto (id e do tipo uuid precisa copiar o codigo id gerado na lista de produtos entre as aspas duplas e colar no campo abaixo)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostra o detalhe do produto com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nome": {
                                                "type": "string",
                                                "example": "Camiseta"
                                            },
                                            "preco": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 49.99
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "atualizar produto",
                "description": "Atualização do produto",
                "operationId": "Atualizar um Produto",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID do Produto (id e do tipo uuid precisa copiar o codigo id gerado na lista de produtos entre as aspas duplas e colar no campo abaixo)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Camiseta do guns"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição da camiseta"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 49.99
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "category_id": {
                                        "type": "string",
                                        "example": "9f2846be-da65-44f6-b453-6abdde55acb1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mostra atualização do produto com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nome": {
                                                "type": "string",
                                                "example": "Camiseta"
                                            },
                                            "preco": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 49.99
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "deletar produto",
                "description": "Ele não deleta o produto inteiro, apenas diminui quantidade e quando essa quantidade chegar a zero ele deleta o produto,\n *     o valor colocado é 1 para demonstrar o funcionamento,mas pode ser alterado dependendo do valor inserido no frontend",
                "operationId": "Deletar um Produto",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID do Produto (id e do tipo uuid precisa copiar o codigo id gerado na lista de produtos entre as aspas duplas e colar no campo abaixo)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostra atualização do produto com sucesso"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}